Fix something

+++ server/src/controllers/cart.Controller.js
@@ -1,14 +1,25 @@
 const asyncHandler = require("../utils/asyncHandler.js");
 const cartService = require("../services/cart.service.js");
+const ErrorResponse = require("../utils/ErrorResponse.js");
+
+const getUserId = (req) => {
+  return req.user?._id || req.user?.id;
+};
 
 const getCart = asyncHandler(async (req, res, next) => {
-  const userId = req.user._id;
+  const userId = getUserId(req);
+  if (!userId) {
+    return next(new ErrorResponse("Not authorized, user ID not found", 401));
+  }
   const result = await cartService.getCartByUserId(userId);
   res.status(200).json(result);
 });
 
 const addToCart = asyncHandler(async (req, res, next) => {
-  const userId = req.user._id;
+  const userId = getUserId(req);
+  if (!userId) {
+    return next(new ErrorResponse("Not authorized, user ID not found", 401));
+  }
   const { foodId, quantity } = req.body;
   const cart = await cartService.addItemToCart(userId, foodId, quantity);
 
@@ -20,7 +31,10 @@ const addToCart = asyncHandler(async (req, res, next) => {
 });
 
 const updateCart = asyncHandler(async (req, res, next) => {
-  const userId = req.user._id;
+  const userId = getUserId(req);
+  if (!userId) {
+    return next(new ErrorResponse("Not authorized, user ID not found", 401));
+  }
   const itemId = req.params.id;
   const { quantity } = req.body;
   
@@ -34,7 +48,10 @@ const updateCart = asyncHandler(async (req, res, next) => {
 });
 
 const removeCartItem = asyncHandler(async (req, res, next) => {
-  const userId = req.user._id;
+  const userId = getUserId(req);
+  if (!userId) {
+    return next(new ErrorResponse("Not authorized, user ID not found", 401));
+  }
   const itemId = req.params.id;
 
   const cart = await cartService.removeItemFromCart(userId, itemId);
